refactor: 도메인 명확화를 위해 Nickname, Point 등 VO 클래스 패키지 이동
