+++ src/main/java/com/refit/api/user/UserController.java
@@ -2,12 +2,12 @@ package com.refit.api.user;
 
 import com.refit.api.user.dto.UserSignupRequest;
 import com.refit.api.user.dto.UserSignupResponse;
+import com.refit.domain.user.entity.type.LoginType;
 import com.refit.domain.user.service.UserService;
 import com.refit.global.api.ApiResponse;
 import lombok.RequiredArgsConstructor;
-import org.springdoc.core.annotations.ParameterObject;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -19,8 +19,9 @@ public class UserController {
     private final UserService userService;
 
     @PostMapping
-    public ApiResponse<UserSignupResponse> localJoin(@ModelAttribute @ParameterObject UserSignupRequest request) {
-        return ApiResponse.success(201, "가입완료", userService.signup(request, null));
+    public ApiResponse<UserSignupResponse> signup(@RequestBody UserSignupRequest request) {
+        UserSignupResponse response = userService.signup(request, LoginType.LOCAL);
+        return ApiResponse.success(201, "회원가입이 완료되었습니다.", response);
     }
 
 }
