+++ src/main/java/com/refit/api/user/dto/UserSignupRequest.java
@@ -15,5 +15,4 @@ public class UserSignupRequest {
     private String email;
     private String password;
     private String confirmPassword;
-    private String nickname;
 }
