+++ src/main/java/com/refit/api/user/dto/UserSignupResponse.java
@@ -8,6 +8,5 @@ import lombok.Setter;
 @Setter
 @Builder
 public class UserSignupResponse {
-    private String email;
     private String nickname;
 }
