--- src/main/java/com/refit/domain/user/entity/Role.java
@@ -1,5 +1,0 @@
-package com.refit.domain.user.entity;
-
-public enum Role {
-    USER, ADMIN
-}
