+++ src/main/java/com/refit/domain/user/entity/type/Role.java
@@ -1,0 +1,5 @@
+package com.refit.domain.user.entity.type;
+
+public enum Role {
+    USER, ADMIN
+}
