+++ src/main/java/com/refit/domain/user/entity/vo/Point.java
@@ -1,4 +1,4 @@
-package com.refit.domain.user.entity;
+package com.refit.domain.user.entity.vo;
 
 import jakarta.persistence.Embeddable;
