+++ src/main/java/com/refit/domain/user/service/UserAuthStrategyResolver.java
@@ -1,13 +1,14 @@
-package com.refit.domain.user.service.signup;
+package com.refit.domain.user.service;
 
-import com.refit.domain.user.entity.LoginType;
+import com.refit.domain.user.entity.type.LoginType;
+import com.refit.domain.user.service.signup.UserSignupService;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class LoginStrategyResolver {
+public class UserAuthStrategyResolver {
 
     private final Map<String, UserSignupService> strategyMap;
